+++ st2common/st2common/services/keyvalues.py
@@ -14,18 +14,57 @@
 # limitations under the License.
 
 from st2common.constants.keyvalue import SYSTEM_SCOPE, USER_SCOPE, ALLOWED_SCOPES
+from st2common.constants.keyvalue import DATASTORE_KEY_SEPARATOR
 from st2common.exceptions.keyvalue import InvalidScopeException, InvalidUserException
 from st2common.models.system.keyvalue import UserKeyReference
 from st2common.persistence.keyvalue import KeyValuePair
 
+__all__ = [
+    'get_kvp_for_name',
+    'get_values_for_names',
+
+    'KeyValueLookup',
+    'UserKeyValueLookup'
+]
+
+
+def get_kvp_for_name(name):
+    try:
+        kvp_db = KeyValuePair.get_by_name(name)
+    except ValueError:
+        kvp_db = None
+
+    return kvp_db
+
+
+def get_values_for_names(names, default_value=None):
+    """
+    Retrieve values for the provided key names (multi get).
+
+    If a KeyValuePair objects for a particular name doesn't exist, the dictionary will contain
+    default_value for that name.
+
+    :rtype: ``dict``
+    """
+    result = {}
+    kvp_dbs = KeyValuePair.get_by_names(names=names)
+
+    name_to_kvp_db_map = {}
+    for kvp_db in kvp_dbs:
+        name_to_kvp_db_map[kvp_db.name] = kvp_db.value
+
+    for name in names:
+        result[name] = name_to_kvp_db_map.get(name, default_value)
+
+    return result
+
 
 class KeyValueLookup(object):
 
-    def __init__(self, key_prefix='', cache=None, scope=SYSTEM_SCOPE):
-        self._key_prefix = key_prefix
-        if cache is None:
-            cache = {}
-        self._value_cache = cache
+    def __init__(self, prefix=None, key_prefix=None, cache=None, scope=SYSTEM_SCOPE):
+        self._prefix = prefix
+        self._key_prefix = key_prefix or ''
+        self._value_cache = cache or {}
         self._scope = scope
 
     def __str__(self):
@@ -39,14 +78,24 @@ class KeyValueLookup(object):
 
     def _get(self, name):
         # get the value for this key and save in value_cache
-        key = '%s.%s' % (self._key_prefix, name) if self._key_prefix else name
-        value = self._get_kv(key)
+        if self._key_prefix:
+            key = '%s.%s' % (self._key_prefix, name)
+        else:
+            key = name
+
+        if self._prefix:
+            kvp_key = DATASTORE_KEY_SEPARATOR.join([self._prefix, key])
+        else:
+            kvp_key = key
+
+        value = self._get_kv(kvp_key)
         self._value_cache[key] = value
         # return a KeyValueLookup as response since the lookup may not be complete e.g. if
         # the lookup is for 'key_base.key_value' it is likely that the calling code, e.g. Jinja,
         # will expect to do a dictionary style lookup for key_base and key_value as subsequent
         # calls. Saving the value in cache avoids extra DB calls.
-        return KeyValueLookup(key_prefix=key, cache=self._value_cache, scope=self._scope)
+        return KeyValueLookup(prefix=self._prefix, key_prefix=key, cache=self._value_cache,
+                              scope=self._scope)
 
     def _get_kv(self, key):
         scope = self._scope
@@ -56,11 +105,10 @@ class KeyValueLookup(object):
 
 class UserKeyValueLookup(object):
 
-    def __init__(self, user, key_prefix='', cache=None, scope=USER_SCOPE):
-        self._key_prefix = key_prefix
-        if cache is None:
-            cache = {}
-        self._value_cache = cache
+    def __init__(self, user, prefix=None, key_prefix=None, cache=None, scope=USER_SCOPE):
+        self._prefix = prefix
+        self._key_prefix = key_prefix or ''
+        self._value_cache = cache or {}
         self._user = user
         self._scope = scope
 
@@ -75,17 +123,24 @@ class UserKeyValueLookup(object):
 
     def _get(self, name):
         # get the value for this key and save in value_cache
-        if not self._key_prefix:
-            key = UserKeyReference(name=name, user=self._user).ref
-        else:
+        if self._key_prefix:
             key = '%s.%s' % (self._key_prefix, name)
-        value = self._get_kv(key)
+        else:
+            key = UserKeyReference(name=name, user=self._user).ref
+
+        if self._prefix:
+            kvp_key = DATASTORE_KEY_SEPARATOR.join([self._prefix, key])
+        else:
+            kvp_key = key
+
+        value = self._get_kv(kvp_key)
         self._value_cache[key] = value
         # return a KeyValueLookup as response since the lookup may not be complete e.g. if
         # the lookup is for 'key_base.key_value' it is likely that the calling code, e.g. Jinja,
         # will expect to do a dictionary style lookup for key_base and key_value as subsequent
         # calls. Saving the value in cache avoids extra DB calls.
-        return UserKeyValueLookup(user=self._user, key_prefix=key, cache=self._value_cache)
+        return UserKeyValueLookup(prefix=self._prefix, user=self._user, key_prefix=key,
+                                  cache=self._value_cache)
 
     def _get_kv(self, key):
         scope = self._scope