+++ st2tests/st2tests/base.py
@@ -33,6 +33,7 @@ from unittest2 import TestCase
 from st2common.exceptions.db import StackStormDBObjectConflictError
 from st2common.models.db import db_setup, db_teardown, db_ensure_indexes
 from st2common.bootstrap.base import ResourceRegistrar
+from st2common.bootstrap.configsregistrar import ConfigsRegistrar
 from st2common.content.utils import get_packs_base_paths
 from st2common.exceptions.db import StackStormDBObjectNotFoundError
 import st2common.models.db.rule as rule_model
@@ -99,6 +100,14 @@ class BaseTestCase(TestCase):
 
         registrar = ResourceRegistrar(use_pack_cache=False)
         registrar.register_packs(base_dirs=get_packs_base_paths())
+
+    @classmethod
+    def _register_pack_configs(self):
+        """
+        Register all the packs inside the fixtures directory.
+        """
+        registrar = ConfigsRegistrar(use_pack_cache=False)
+        registrar.register_configs_for_all_packs(base_dirs=get_packs_base_paths())
 
 
 class EventletTestCase(TestCase):
@@ -186,6 +195,7 @@ class DbTestCase(BaseDbTestCase):
     db_connection = None
     current_result = None
     register_packs = False
+    register_pack_configs = False
 
     @classmethod
     def setUpClass(cls):
@@ -194,6 +204,9 @@ class DbTestCase(BaseDbTestCase):
 
         if cls.register_packs:
             cls._register_packs()
+
+        if cls.register_pack_configs:
+            cls._register_pack_configs()
 
     @classmethod
     def tearDownClass(cls):
@@ -297,8 +310,17 @@ class CleanDbTestCase(BaseDbTestCase):
     database.
     """
 
+    register_packs = False
+    register_pack_configs = False
+
     def setUp(self):
         self._establish_connection_and_re_create_db()
+
+        if self.register_packs:
+            self._register_packs()
+
+        if self.register_pack_configs:
+            self._register_pack_configs()
 
 
 class CleanFilesTestCase(TestCase):