+++ CHANGELOG.rst
@@ -27,8 +27,6 @@ in development
 * Add ``get_fixture_content`` method to all the base pack resource test classes. This method
   enforces fixture files location and allows user to load raw fixture content from a file on disk.
   (new feature)
-* Introduce a new concept of pack config schemas. Each pack can now contain a
-  ``config.schema.yaml`` file. This file can contain an optional schema for the pack config. In the
   future, pack configs will be validated against the schema (if available). (new feature)
 * Add data model and API changes for supporting user scoped variables. (new-feature, experimental)
 * Add missing `pytz` dependency to ``st2client`` requirements file. (bug-fix)
@@ -38,8 +36,16 @@ in development
   JSON has been deprecated in StackStorm v0.6. Now the only supported metadata file format is YAML.
 * Add ``-y`` / ``--yaml`` flag to the CLI ``list`` and ``get`` commands. If this flag is provided,
   command response will be formatted as YAML. (new feature)
-* Alias names are scoped to a pack. (bug-fix)
+* Alias names are now correctly scoped to a pack. This means the same name for alias can be used
+  across different packs. (bug-fix)
 * Ability to migrate api keys to new installs. (new feature)
+* Introduce a new concept of pack config schemas. Each pack can now contain a
+  ``config.schema.yaml`` file. This file can contain an optional schema for the pack config. In the
+* Introduce support for pack configs which are located outside of the pack directory in
+  ``/opt/stackstorm/configs/<pack name>.yaml`` files. Those files are similar to the existing pack
+  configs, but in addition to the static values they can also contain dynamic values. Dynamic value
+  is a value which contains a Jinja expression which is resolved to the datastore item during
+  run-time. (new feature)
 
 1.4.0 - April 18, 2016
 ----------------------